requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.5
django-bootstrap4==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
phonenumberslite==8.13.21
psycopg2==2.9.7
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
